fix issues on added shows

+++ DBase/DTos/ShowOpening.cs
@@ -1,6 +1,6 @@
 public class ShowOpening
 {
-    public int ShowId { get; set; }
+    public long ShowId { get; set; }
     public DateTime StartDate { get; set; }
     public DateTime EndDate { get; set; }
 }
